# openapi.yaml
openapi: 3.0.3
info:
  title: YieldScan Backend API
  description: |
    API for managing crypto assets, their yield-generating steps, and retrieving APY data.
    This API supports protocol-based step organization with comprehensive asset and step management capabilities.
  version: 1.0.0
  contact:
    name: YieldScan Team
servers:
  - url: http://localhost:4023/api
    description: Local development server
  - url: https://api.yieldscan.com/api
    description: Production server
paths:
...
  /assets:
    get:
      tags:
        - Assets
      summary: Get all assets
      description: Retrieve all assets with optional filtering
      parameters:
        - name: chain
          in: query
          description: Filter by blockchain network
          schema:
            type: string
            example: ETH
        - name: protocol
          in: query
          description: Filter by protocol name
          schema:
            type: string
            example: Aave
        - name: yieldBearingToken
          in: query
          description: Filter by yield-bearing token status
          schema:
            type: boolean
        - name: token
          in: query
          description: Filter by token name (case-insensitive regex)
          schema:
            type: string
            example: USDT
        - name: enabled
          in: query
          description: Filter by enabled status
          schema:
            type: boolean
        - name: includeDisabled
          in: query
          description: Include disabled assets in results
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'
                  count:
                    type: integer
                    example: 25
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Assets
      summary: Create a new asset
      description: Create a new crypto asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Asset created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Asset created successfully
                  asset:
                    $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /chains:
    get:
      tags:
        - Chains
      summary: Get all chains
      description: Retrieve all blockchain networks with optional filtering
      parameters:
        - name: enabled
          in: query
          description: Filter by enabled status
          schema:
            type: boolean
        - name: testnet
          in: query
          description: Filter by testnet status
          schema:
            type: boolean
        - name: includeDisabled
          in: query
          description: Include disabled chains in results
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  chains:
                    type: array
                    items:
                      $ref: '#/components/schemas/Chain'
                  count:
                    type: integer
                    example: 3
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Chains
      summary: Create a new chain
      description: Create a new blockchain network configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChainRequest'
      responses:
        '201':
          description: Chain created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Chain created successfully
                  chain:
                    $ref: '#/components/schemas/Chain'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /protocols:
    get:
      tags:
        - Protocols
      summary: Get all protocols
      description: Retrieve all DeFi protocols with optional filtering
      parameters:
        - name: enabled
          in: query
          description: Filter by enabled status
          schema:
            type: boolean
        - name: category
          in: query
          description: Filter by protocol category
          schema:
            type: string
            example: Lending
        - name: includeDisabled
          in: query
          description: Include disabled protocols in results
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  protocols:
                    type: array
                    items:
                      $ref: '#/components/schemas/Protocol'
                  count:
                    type: integer
                    example: 5
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Protocols
      summary: Create a new protocol
      description: Create a new DeFi protocol configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProtocolRequest'
      responses:
        '201':
          description: Protocol created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Protocol created successfully
                  protocol:
                    $ref: '#/components/schemas/Protocol'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /apy:
    get:
      tags:
        - APY Data
      summary: Get APY data
      description: Retrieve current APY data for all supported tokens across different protocols
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  description: Chain ID as key
                  additionalProperties:
                    type: object
                    description: Token address as key
                    properties:
                      compound:
                        type: number
                        nullable: true
                        example: 4.25
                      aave:
                        type: number
                        nullable: true
                        example: 3.87
                      radiant:
                        type: number
                        nullable: true
                        example: 5.12
                      venus:
                        type: number
                        nullable: true
                        example: 6.33
              example:
                "1":
                  "0xdAC17F958D2ee523a2206206994597C13D831ec7":
                    compound: 4.25
                    aave: 3.87
                    radiant: null
                    venus: null
                "56":
                  "0x55d398326f99059ff775485246999027b3197955":
                    compound: null
                    aave: null
                    radiant: 2.15
                    venus: 6.33
        '500':
          $ref: '#/components/responses/InternalServerError'

  # User Details API
  /user-details/{walletAddress}:
    get:
      tags:
        - User Details
      summary: Get user transaction details
      description: Retrieve transaction history and details for a specific wallet address
      parameters:
        - name: walletAddress
          in: path
          required: true
          description: Wallet address (case-insensitive)
          schema:
            type: string
            example: "0x742d35Cc6639C0532fE47bD00E08C7D7c4B0d62f"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                description: User transaction data (structure varies based on implementation)
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    Asset:
      type: object
      properties:
        _id:
          type: string
          example: 684a8cf6db44b98888f22f5f
        token:
          type: string
          example: USDT
        chain:
          type: string
          example: ETH
        address:
          type: string
          example: "0xdAC17F958D2ee523a2206206994597C13D831ec7"
        icon:
          type: string
          example: "https://cdn.jsdelivr.net/gh/trustwallet/assets/blockchains/ethereum/assets/0xdAC17F958D2ee523a2206206994597C13D831ec7/logo.png"
        chainId:
          type: integer
          example: 1
        decimals:
          type: integer
          example: 6
        maxDecimalsShow:
          type: integer
          example: 2
        usdPrice:
          type: number
          example: 1.0
        yieldBearingToken:
          type: boolean
          example: false
        enabled:
          type: boolean
          example: true
        underlyingAsset:
          type: string
          nullable: true
          example: null
        withdrawContract:
          type: string
          nullable: true
          example: null
        protocol:
          type: string
          nullable: true
          example: null
        withdrawUri:
          type: string
          nullable: true
          example: null
        maturity:
          type: string
          nullable: true
          example: null
        lockYield:
          $ref: '#/components/schemas/LockYield'
        createdAt:
          type: string
          format: date-time
          example: "2025-06-16T08:11:29.443Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-06-16T08:25:35.229Z"
    CreateAssetRequest:
      type: object
      properties:
        token:
          type: string
          example: USDT
        chain:
          type: string
          example: ETH
        address:
          type: string
          example: "0xdAC17F958D2ee523a2206206994597C13D831ec7"
        icon:
          type: string
          example: "https://cdn.jsdelivr.net/gh/trustwallet/assets/blockchains/ethereum/assets/0xdAC17F958D2ee523a2206206994597C13D831ec7/logo.png"
        chainId:
          type: integer
          example: 1
        decimals:
          type: integer
          example: 6
        maxDecimalsShow:
          type: integer
          example: 2
        usdPrice:
          type: number
          example: 1.0
        yieldBearingToken:
          type: boolean
          example: false
        enabled:
          type: boolean
          example: true
        underlyingAsset:
          type: string
          nullable: true
        withdrawContract:
          type: string
          nullable: true
        protocol:
          type: string
          nullable: true
        withdrawUri:
          type: string
          nullable: true
        maturity:
          type: string
          nullable: true
        lockYield:
          $ref: '#/components/schemas/LockYield'
      required:
        - token
        - chain
        - address
        - icon
        - chainId
        - decimals
        - maxDecimalsShow
        - usdPrice
        - yieldBearingToken
...
    Chain:
      type: object
      properties:
        _id:
          type: string
          example: 684a8cf6db44b98888f22f5f
        chainId:
          type: integer
          example: 1
        name:
          type: string
          example: Ethereum
        shortName:
          type: string
          example: ETH
        image:
          type: string
          example: "https://cdn.jsdelivr.net/gh/trustwallet/assets/blockchains/ethereum/info/logo.png"
        rpcUrl:
          type: string
          example: "https://eth.llamarpc.com"
        explorerUrl:
          type: string
          example: "https://etherscan.io"
        symbol:
          type: string
          example: ETH
        decimals:
          type: integer
          example: 18
        enabled:
          type: boolean
          example: true
        testnet:
          type: boolean
          nullable: true
          example: false
        description:
          type: string
          nullable: true
          example: "Ethereum mainnet"
        createdAt:
          type: string
          format: date-time
          example: "2025-06-16T08:11:29.443Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-06-16T08:25:35.229Z"
    CreateChainRequest:
      type: object
      properties:
        chainId:
          type: integer
          example: 1
        name:
          type: string
          example: Ethereum
        shortName:
          type: string
          example: ETH
        image:
          type: string
          example: "https://cdn.jsdelivr.net/gh/trustwallet/assets/blockchains/ethereum/info/logo.png"
        rpcUrl:
          type: string
          example: "https://eth.llamarpc.com"
        explorerUrl:
          type: string
          example: "https://etherscan.io"
        symbol:
          type: string
          example: ETH
        decimals:
          type: integer
          example: 18
        enabled:
          type: boolean
          example: true
        testnet:
          type: boolean
          nullable: true
          example: false
        description:
          type: string
          nullable: true
          example: "Ethereum mainnet"
      required:
        - chainId
        - name
        - shortName
        - image
        - rpcUrl
        - explorerUrl
        - symbol
        - decimals
...
    Protocol:
      type: object
      properties:
        _id:
          type: string
          example: 684a8cf6db44b98888f22f5f
        name:
          type: string
          example: Aave
        displayName:
          type: string
          example: Aave
        image:
          type: string
          example: "https://cdn.jsdelivr.net/gh/trustwallet/assets/blockchains/ethereum/assets/0x7Fc66500c84A76Ad7e9c93437bFc5Ac33E2DDaE9/logo.png"
        website:
          type: string
          nullable: true
          example: "https://aave.com"
        description:
          type: string
          nullable: true
          example: "Open source and non-custodial liquidity protocol for earning interest on deposits and borrowing assets"
        category:
          type: string
          nullable: true
          example: "Lending"
        enabled:
          type: boolean
          example: true
        supportedChains:
          type: array
          items:
            type: integer
          nullable: true
          example: [1, 137, 42161, 8453, 56]
        createdAt:
          type: string
          format: date-time
          example: "2025-06-16T08:11:29.443Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-06-16T08:25:35.229Z"
    CreateProtocolRequest:
      type: object
      properties:
        name:
          type: string
          example: Aave
        displayName:
          type: string
          example: Aave
        image:
          type: string
          example: "https://cdn.jsdelivr.net/gh/trustwallet/assets/blockchains/ethereum/assets/0x7Fc66500c84A76Ad7e9c93437bFc5Ac33E2DDaE9/logo.png"
        website:
          type: string
          nullable: true
          example: "https://aave.com"
        description:
          type: string
          nullable: true
          example: "Open source and non-custodial liquidity protocol for earning interest on deposits and borrowing assets"
        category:
          type: string
          nullable: true
          example: "Lending"
        enabled:
          type: boolean
          example: true
        supportedChains:
          type: array
          items:
            type: integer
          nullable: true
          example: [1, 137, 42161, 8453, 56]
      required:
        - name
        - displayName
        - image
...
    LockYield:
      type: object
      nullable: true
      properties:
        expirationDate:
          type: string
          example: "2024-12-31"
        protocol:
          $ref: '#/components/schemas/LockYieldProtocol'
    LockYieldProtocol:
      type: object
      properties:
        name:
          type: string
          example: "Pendle"
        swap:
          type: boolean
          example: true
        ytAddress:
          type: string
          example: "0x..."
        ptAddress:
          type: string
          example: "0x..."
        swapAddress:
          type: string
          example: "0x..."
        ytMarketAddress:
          type: string
          example: "0x..."
        ytDecimals:
          type: integer
          example: 18
        ptDecimals:
          type: integer
          example: 18
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        details:
          type: string
          example: "Missing required field: token"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Missing required fields"
            details: "assetId, order, title, description, chain, functionConfig are required"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Failed to fetch assets"
            details: "Database connection error"
...